#20230003 (#43)
* Docker update

Worked on making the script work on a docker image and is now compatible with AWS Lightsail and DB

* Experimental changes to created docker version

* Changes made with David

Started a complete recode of backend to try to pinpoint dysnc issue. Still having it prob something stupid. Work on backend testing to make sure we have it all working together seamlessly

* Fixed de-sync error

Best version of backend that now compensates for 2 seconds of de-sync properly

* Update to update user info if missing and moved to Melly DB

* Updating loaders, cleaning up for when we push

Still need to deprecate old code

* Sync with laptop

* Added basic logging of user information

* Removed Old Security code, Cleaning up Repo

* Fixed bad variable name, encryption now works the same for heartbeat and regular testing

* Made updates to try to get it to work on Lightsail

* Updated misc file fixes